requirements-local.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Audio Processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1

# Machine Learning
torch==2.1.1
torchaudio==2.1.1
openai-whisper==20231117
transformers==4.35.2

# Optional: Speaker Diarization
pyannote.audio==3.1.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 